requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
gunicorn